+++ CHANGELOG.md
@@ -1,3 +1,5 @@
+* Fixed closing of child driver with shared balancer
+
 ## v3.89.6
 * Refactored `database/sql` driver internals for query-service client support in the future